fix: 유형자산 비유형자산 잘못 매칭되어서 수정
